Fix #663, Scrub datatypes across ES
Fix all use of uint32 for sizes, replace with size_t
Scrub all other types to make sure use is consistent,
using proper typedef whenever available.

+++ fsw/cfe-core/src/es/cfe_es_perf.c
@@ -274,7 +274,7 @@ bool CFE_ES_RunPerfLogDump(uint32 ElapsedTime, void *Arg)
     CFE_ES_PerfDumpGlobal_t *State = (CFE_ES_PerfDumpGlobal_t *)Arg;
     int32               Status;
     CFE_FS_Header_t     FileHdr;
-    uint32              BlockSize;
+    size_t              BlockSize;
 
     /*
      * each time this background job is re-entered after a time delay,
